move EcalPedestalHistory in to plugins directory to avoid duplicate definitions
